# Performance fixes for script interpreter