--- a/Engine/script/cc_instance.cpp
+++ b/Engine/script/cc_instance.cpp
@@ -480,13 +480,15 @@ int ccInstance::Run(int32_t curpc)
     const auto timeout_abort = std::chrono::milliseconds(_timeoutAbortMs);
     _lastAliveTs = AGS_FastClock::now();
 
-    while ((flags & INSTF_ABORTED) == 0) {
-        /*
-		if (!codeInst->ReadOperation(codeOp, pc))
-        {
-            return -1;
-        }
-        */
+    /* Main bytecode execution loop */
+    //=====================================================================
+    while ((flags & INSTF_ABORTED) == 0)
+    {
+        // WARNING: a time-critical code ahead;
+        // trying to pick some of the code out to separate function(s)
+        // may lead to a performance loss in script-heavy games.
+        // always compare execution speed before applying any major changes!
+        //
         /* ReadOperation */
         //=====================================================================
         codeOp.Instruction.Code			= codeInst->code[pc];
@@ -512,15 +514,9 @@ int ccInstance::Run(int32_t curpc)
             char fixup = codeInst->code_fixups[pc_at];
             if (fixup > 0)
             {
-                // could be relative pointer or import address
-                /*
-                if (!FixupArgument(code[pc], fixup, codeOp.Args[i]))
-                {
-                    return -1;
-                }
-                */
                 /* FixupArgument */
                 //=====================================================================
+                // could be relative pointer or import address
                 switch (fixup)
                 {
                 case FIXUP_GLOBALDATA:
@@ -588,7 +584,8 @@ int ccInstance::Run(int32_t curpc)
             DumpInstruction(codeOp);
         }
 
-        switch (codeOp.Instruction.Code) {
+        switch (codeOp.Instruction.Code)
+        {
       case SCMD_LINENUM:
           line_number = arg1.IValue;
           currentline = arg1.IValue;
@@ -1887,89 +1884,6 @@ bool ccInstance::ResolveImportFixups(const ccScript *scri)
     }
     return true;
 }
-
-/*
-bool ccInstance::ReadOperation(ScriptOperation &op, int32_t at_pc)
-{
-	op.Instruction.Code			= code[at_pc];
-	op.Instruction.InstanceId	= (op.Instruction.Code >> INSTANCE_ID_SHIFT) & INSTANCE_ID_MASK;
-	op.Instruction.Code		   &= INSTANCE_ID_REMOVEMASK; // now this is pure instruction code
-
-    int want_args = sccmd_info[op.Instruction.Code].ArgCount;
-    if (at_pc + want_args >= codesize)
-    {
-        cc_error("unexpected end of code data at %d", at_pc + want_args);
-        return false;
-    }
-    op.ArgCount = want_args;
-
-    at_pc++;
-    for (int i = 0; i < op.ArgCount; ++i, ++at_pc)
-    {
-        char fixup = code_fixups[at_pc];
-        if (fixup > 0)
-        {
-            // could be relative pointer or import address
-            if (!FixupArgument(code[at_pc], fixup, op.Args[i]))
-            {
-                return false;
-            }
-        }
-        else
-        {
-            // should be a numeric literal (int32 or float)
-            op.Args[i].SetInt32( (int32_t)code[at_pc] );
-        }
-    }
-
-    return true;
-}
-*/
-/*
-bool ccInstance::FixupArgument(intptr_t code_value, char fixup_type, RuntimeScriptValue &argument)
-{
-    switch (fixup_type)
-    {
-    case FIXUP_GLOBALDATA:
-        {
-            ScriptVariable *gl_var = (ScriptVariable*)code_value;
-            argument.SetGlobalVar(&gl_var->RValue);
-        }
-        break;
-    case FIXUP_FUNCTION:
-        // originally commented -- CHECKME: could this be used in very old versions of AGS?
-        //      code[fixup] += (long)&code[0];
-        // This is a program counter value, presumably will be used as SCMD_CALL argument
-        argument.SetInt32((int32_t)code_value);
-        break;
-    case FIXUP_STRING:
-        argument.SetStringLiteral(&strings[0] + code_value);
-        break;
-    case FIXUP_IMPORT:
-        {
-            const ScriptImport *import = simp.getByIndex((int32_t)code_value);
-            if (import)
-            {
-                argument = import->Value;
-            }
-            else
-            {
-                cc_error("cannot resolve import, key = %ld", code_value);
-                return false;
-            }
-        }
-        break;
-    case FIXUP_STACK:
-        argument = GetStackPtrOffsetFw((int32_t)code_value);
-        break;
-    default:
-        cc_error("internal fixup type error: %d", fixup_type);
-        return false;;
-    }
-    return true;
-}
-*/
-//-----------------------------------------------------------------------------
 
 void ccInstance::PushValueToStack(const RuntimeScriptValue &rval)
 {