Resolves #843 and #1987.

Contains a number of adjustments to the script interpreter function (ccInstance::Run), increases potential game fps and reduces CPU load for "script heavy" games.

Adds `DEBUG_CC_EXEC` compilation flag, and hides most error checks within the opcode implementations under this flag.
DEBUG_CC_EXEC is enabled if AGS_PLATFORM_DEBUG is enabled, but we may also make it completely separate later if wanted.

The changes were tested against two games with complex scripts in the "infinite fps" mode (where engine runs game loops as fast as it can):
1. First person space sim prototype: https://github.com/ivan-mogilko/ags-wcproto
2. Car racing arcade game: https://github.com/ivan-mogilko/ags-lastfurious

First has generally very slow script; showed a fps rise from ~64-65 to ~75 in idle state.
Second - from ~145-155 fps to 180-190 fps, when idling at the starting grid while 5 ai cars were racing the loop.
which gives an improvement by something like 15-20% depending on a case.
**UPDATE:** after few more commits made since:
First game: rise from ~65 to 80 fps
Second game: rise from 150 to 200 fps
**UPDATE:** after another [sudden idea](https://github.com/adventuregamestudio/ags/pull/1990/commits/719f67e831f7de3da831a957da0cc69a7a4a5026):
First game: rise from ~65 to 84 fps
Second game: rise from 150 to 210 fps
So, this is improvement by ~30% now, or more, depending on a game.


These numbers are of course subjective to an individual system (and game), but probably may give an idea of improvement.