@ericoporto 2023-04-28
> but we may also make it completely separate later if wanted.

In cc_instance.h, use

```
#ifndef DEBUG_CC_EXEC
#define DEBUG_CC_EXEC (AGS_PLATFORM_DEBUG)
#endif
```