@ericoporto 2023-04-29
Along with the commit https://github.com/adventuregamestudio/ags/pull/1990/commits/d9284d8c1a265f0c4b329c27dde2997f465375d4

I suggest making the loopinteractions check use 1024 instead of 1000 too (https://github.com/ericoporto/ags/commit/513d18daf797b13492d4a11893f8ef8aed7790e3)