@ivan-mogilko 2023-04-29
So, after a curious @ericoporto 's idea to [replace a argument fixup loop with a switch](https://github.com/adventuregamestudio/ags/pull/1990/commits/a495a00f08c9aed395d6ad6513dab35d551fd613), which gave another good boost to performance at a cost of somewhat ugly code, I suddenly had a thought that we were doing this wrong all this time. Analysing the fixup uses shows that they potentially applied only to literal args, and only when arg can be a "address" (to a script, offset in stack, pointer to function, pointer to a object, etc). In the end this narrowed down to only *two* instructions: SCMD_WRITELIT and SCMD_LITTOREG.

*Unless I am making some stupid mistake*, these two instructions SCMD_WRITELIT and SCMD_LITTOREG are the only ones which argument may potentially be a result of fixup. So I pushed [another commit](https://github.com/adventuregamestudio/ags/pull/1990/commits/719f67e831f7de3da831a957da0cc69a7a4a5026) which simply deletes a fixup loop/switch in front, and only calls FixupArgument explicitly for these two instructions.

This tidied the code considerably, and also gave little more execution speed. I tested couple of games, and so far there seem to be no script errors, so maybe I'm right in my assumption.