@ivan-mogilko 2023-05-01
Merged!

This PR results were unexpected, I've compared the latest version with the wcproto ported to 3.2.1 (I might push that to my game's repository later), and the fps is pretty close to 3.2.1 engine (maybe 5-7% worse). This is a massive improvement compared to where we were at 3.6.0 release.